fastapi-ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.3
pydantic==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.25.2
aiofiles==23.2.1
google-generativeai==0.3.1
